+++ packages/core/src/core/transformData.ts
@@ -1,5 +1,5 @@
 import { options } from './options';
-import { fromHttpStatus, interceptStr, getTimestamp, parseUrlParams } from '@zmonitor/utils';
+import { fromHttpStatus, interceptStr, getTimestamp, variableTypeDetection } from '@zmonitor/utils';
 import { HTTP_CODE, STATUS_CODE } from '@zmonitor/common';
 import { HttpData, ResouceError, ResourceTarget } from '@zmonitor/types';
 
@@ -9,6 +9,30 @@ function safeStringify(response: any): string {
   } catch (e) {
     return '无法序列化的响应';
   }
+}
+
+// 解析url参数
+export function parseUrlParams(url: string): { [key: string]: string } {
+  // 定义一个对象来存储参数
+  const params: { [key: string]: string } = {};
+
+  // 检查传入的url是否为字符串
+  if (variableTypeDetection.isString(url)) {
+    // 获取查询字符串部分
+    const queryString = url.split('?')[1];
+    if (queryString) {
+      // 分割查询字符串
+      const pairs = queryString.split('&');
+      pairs.forEach(pair => {
+        // 分割键值对
+        const [key, value] = pair.split('=');
+        // 解码并存储在params对象中
+        params[decodeURIComponent(key)] = decodeURIComponent(value || '');
+      });
+    }
+  }
+
+  return params;
 }
 
 // 处理接口的状态
@@ -55,7 +79,8 @@ export function httpTransform(data: HttpData): HttpData {
     },
     response: {
       Status,
-      data: status == STATUS_CODE.ERROR ? response : null,
+      data: response,
+      // data: status == STATUS_CODE.ERROR ? response : null,
     },
   };
 }