+++ packages/recordscreen/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @zmonitor/recordscreen
 
+## 4.0.0
+
+### Major Changes
+
+- test
+
+### Patch Changes
+
+- Updated dependencies
+  - @zmonitor/common@4.0.0
+  - @zmonitor/types@4.0.0
+  - @zmonitor/utils@4.0.0
+
 ## 3.0.0
 
 ### Major Changes